requirements.txt
aiogram>=3.4
anthropic>=0.25
openpyxl>=3.1
python-docx>=1.1
pdf2image>=1.17
aiohttp>=3.9
aiofiles>=23.2
python-dotenv>=1.0
